Makes grid based on width and height props
These are just defaulted to 8 now. Someday, however,
this will be a dynamic input. This is being done to
help fetching a square's neighbor plants a bit more
simple.

Note: though this is now dynamic the current rendering
pattern encounters serious performance issues when more
than about 120 squares are in place