Allows canDrop unless already has plant
The point of the app is for users to have data. If they
know something the simple algorithm does _or_ want to
put a plant in a bad decision they should be able to.

This also avoids an odd case (perhaps a bug) in the
renderOverlay checking of GardenSquare#render