Styles recommendations based on rating
Formerly a square was red or green based on whether it
has a friend or enemy in the neighborhood. This commit
replaces "hasOverlap" with "getRating" which returns
a relative score based on the number of relationships
and the distance of those relators.

This rating is used to determine the overlay's opacity.

This commit also switches up the look and feel a bit
removing dark and temporarily putting in place an svg
from the cssgridgarden.com to make the app look at little
less like a game of chess!