Adds localStorage fallback
Until now right after logging in the auth_token was not being sent
as a header in the fetch call to GET /gardens until a page was reloaded

This is an indicatory that this whole Auth concept is a bit janky. Each
copy of it is now mutated when a login or signup happens. As such this
is a clear signal that this needs to be moved into something like
redux state.