Refactors auth to use UserContext
This has the primary benefit of isolating all this app/auth/user to a
single file inside the UserProvider component to be more specific.

It's the hope that this will help avoid the need for passing around
semi-useless components (like authenticated was previously) and overtime
help reduce nesting.

Might be refactored again in future to be hooks. Whatever's cool, right?